Rename DOF to Joint viewer
