+++ exporters/InventorRobotExporter/InventorRobotExporter/GUI/Guide/Guide.html
@@ -47,7 +47,7 @@
             <li>Sensor setup</li>
         </ol>
     </li>
-    <li>View the degrees of freedom using the <i style="font-weight: bold">Toggle Degrees of Freedom View</i> button.
+    <li>View the status of robot joints using the <i style="font-weight: bold">Toggle Joint Viewer</i> button.
     </li>
 </ol>
 </body>
