+++ exporters/InventorRobotExporter/InventorRobotExporter/GUI/JointView/JointViewKeyPane.cs
@@ -1,10 +1,10 @@
 ﻿using System.Windows.Forms;
 
-namespace InventorRobotExporter.GUI.DegreesOfFreedomViewer
+namespace InventorRobotExporter.GUI.JointView
 {
-    public partial class DofKeyPane : UserControl
+    public partial class JointViewKeyPane : UserControl
     {
-        public DofKeyPane()
+        public JointViewKeyPane()
         {
             InitializeComponent();
             webBrowser1.DocumentText = Properties.Resources.DOFKey;
