+++ exporters/InventorRobotExporter/InventorRobotExporter/GUI/JointView/JointViewKeyPane.resx
